requirements.txt
numpy==1.26.0
pandas==2.1.1
pytorch-lightning==2.1.0
pyyaml==6.0.1
scikit-learn==1.3.2
tensorboard==2.15.0
torch==2.1.0
tqdm==4.65.0
yaspin==3.0.1
